+++ packages/create-app/templates/vue-ts-static/vite.config.ts
@@ -1,0 +1,8 @@
+import vue from '@vitejs/plugin-vue'
+import { defineConfig } from 'vite'
+import { createBlockletPlugin } from 'vite-plugin-blocklet'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [vue(), createBlockletPlugin()],
+})